requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
tensorflow>=2.12.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0

# NLP & Transformers
transformers>=4.25.0
datasets>=2.8.0
tokenizers>=0.13.0
safetensors>=0.4.0

# Security Analysis
radare2>=5.8.0
r2pipe>=1.7.0
capstone>=5.0.0
ropper>=1.13.0
angr>=9.2.0
pefile>=2023.2.0
lief>=0.13.0

# Web & Network
requests>=2.28.0
beautifulsoup4>=4.11.0
scapy>=2.4.0
urllib3>=1.26.0

# Code Analysis
tree-sitter>=0.20.0
pycparser>=2.21.0
javalang>=0.13.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
wandb>=0.15.0
